requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-crispy-forms==1.14.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
